+++ src/main/java/net/neoforged/neoforge/internal/NeoForgeStatesProvider.java
@@ -11,6 +11,7 @@ import net.neoforged.fml.IModStateProvider;
 import net.neoforged.fml.ModLoadingPhase;
 import net.neoforged.fml.ModLoadingState;
 import net.neoforged.neoforge.common.capabilities.CapabilityManager;
+import net.neoforged.neoforge.common.world.chunk.ForcedChunkManager;
 import net.neoforged.neoforge.network.NetworkRegistry;
 import net.neoforged.neoforge.registries.GameData;
 import net.neoforged.neoforge.registries.RegistryManager;
@@ -21,10 +22,11 @@ public class NeoForgeStatesProvider implements IModStateProvider {
     final ModLoadingState UNFREEZE = ModLoadingState.withInline("UNFREEZE_DATA", "INJECT_CAPABILITIES", ModLoadingPhase.GATHER, ml -> GameData.unfreezeData());
     final ModLoadingState LOAD_REGISTRIES = ModLoadingState.withInline("LOAD_REGISTRIES", "UNFREEZE_DATA", ModLoadingPhase.GATHER, ml -> GameData.postRegisterEvents());
     final ModLoadingState FREEZE = ModLoadingState.withInline("FREEZE_DATA", "LOAD_REGISTRIES", ModLoadingPhase.GATHER, ml -> GameData.freezeData());
+    final ModLoadingState REGISTER_TICKET_CONTROLLERS = ModLoadingState.withInline("REGISTER_TICKET_CONTROLLERS", "COMPLETE", ModLoadingPhase.COMPLETE, ml -> ForcedChunkManager.init());
     final ModLoadingState NETLOCK = ModLoadingState.withInline("NETWORK_LOCK", "COMPLETE", ModLoadingPhase.COMPLETE, ml -> NetworkRegistry.lock());
 
     @Override
     public List<IModLoadingState> getAllStates() {
-        return List.of(CREATE_REGISTRIES, INJECT_CAPABILITIES, UNFREEZE, LOAD_REGISTRIES, FREEZE, NETLOCK);
+        return List.of(CREATE_REGISTRIES, INJECT_CAPABILITIES, UNFREEZE, LOAD_REGISTRIES, FREEZE, REGISTER_TICKET_CONTROLLERS, NETLOCK);
     }
 }
