+++ tests/src/main/java/net/neoforged/neoforge/debug/world/ForgeChunkManagerTest.java
@@ -5,11 +5,10 @@
 
 package net.neoforged.neoforge.debug.world;
 
-import com.mojang.datafixers.util.Pair;
-import it.unimi.dsi.fastutil.longs.LongSet;
 import java.util.Map;
 import java.util.UUID;
 import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.CreativeModeTabs;
@@ -21,9 +20,10 @@ import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.MapColor;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
-import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.neoforged.neoforge.common.world.ForcedChunkManager;
+import net.neoforged.neoforge.common.world.chunk.RegisterTicketControllersEvent;
+import net.neoforged.neoforge.common.world.chunk.TicketController;
+import net.neoforged.neoforge.common.world.chunk.TicketSet;
 import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
 import net.neoforged.neoforge.registries.DeferredBlock;
 import net.neoforged.neoforge.registries.DeferredItem;
@@ -39,10 +39,29 @@ public class ForgeChunkManagerTest {
     private static final DeferredRegister.Items ITEMS = DeferredRegister.createItems(MODID);
     private static final DeferredBlock<Block> CHUNK_LOADER_BLOCK = BLOCKS.register("chunk_loader", () -> new ChunkLoaderBlock(Properties.of().mapColor(MapColor.STONE)));
     private static final DeferredItem<BlockItem> CHUNK_LOADER_ITEM = ITEMS.registerBlockItem(CHUNK_LOADER_BLOCK);
+    private static final TicketController CONTROLLER = new TicketController(new ResourceLocation(MODID, "default"), (world, ticketHelper) -> {
+        for (Map.Entry<BlockPos, TicketSet> entry : ticketHelper.getBlockTickets().entrySet()) {
+            BlockPos key = entry.getKey();
+            int ticketCount = entry.getValue().nonTicking().size();
+            int tickingTicketCount = entry.getValue().ticking().size();
+            if (world.getBlockState(key).is(CHUNK_LOADER_BLOCK.get()))
+                LOGGER.info("Allowing {} chunk tickets and {} ticking chunk tickets to be reinstated for position: {}.", ticketCount, tickingTicketCount, key);
+            else {
+                ticketHelper.removeAllTickets(key);
+                LOGGER.info("Removing {} chunk tickets and {} ticking chunk tickets for no longer valid position: {}.", ticketCount, tickingTicketCount, key);
+            }
+        }
+        for (Map.Entry<UUID, TicketSet> entry : ticketHelper.getEntityTickets().entrySet()) {
+            UUID key = entry.getKey();
+            int ticketCount = entry.getValue().nonTicking().size();
+            int tickingTicketCount = entry.getValue().ticking().size();
+            LOGGER.info("Allowing {} chunk tickets and {} ticking chunk tickets to be reinstated for entity: {}.", ticketCount, tickingTicketCount, key);
+        }
+    });
 
     public ForgeChunkManagerTest() {
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-        modEventBus.addListener(this::commonSetup);
+        modEventBus.addListener(this::registerControllers);
         BLOCKS.register(modEventBus);
         ITEMS.register(modEventBus);
         modEventBus.addListener(this::addCreative);
@@ -53,26 +72,8 @@ public class ForgeChunkManagerTest {
             event.accept(CHUNK_LOADER_ITEM);
     }
 
-    private void commonSetup(FMLCommonSetupEvent event) {
-        event.enqueueWork(() -> ForcedChunkManager.setForcedChunkLoadingCallback(MODID, (world, ticketHelper) -> {
-            for (Map.Entry<BlockPos, Pair<LongSet, LongSet>> entry : ticketHelper.getBlockTickets().entrySet()) {
-                BlockPos key = entry.getKey();
-                int ticketCount = entry.getValue().getFirst().size();
-                int tickingTicketCount = entry.getValue().getSecond().size();
-                if (world.getBlockState(key).is(CHUNK_LOADER_BLOCK.get()))
-                    LOGGER.info("Allowing {} chunk tickets and {} ticking chunk tickets to be reinstated for position: {}.", ticketCount, tickingTicketCount, key);
-                else {
-                    ticketHelper.removeAllTickets(key);
-                    LOGGER.info("Removing {} chunk tickets and {} ticking chunk tickets for no longer valid position: {}.", ticketCount, tickingTicketCount, key);
-                }
-            }
-            for (Map.Entry<UUID, Pair<LongSet, LongSet>> entry : ticketHelper.getEntityTickets().entrySet()) {
-                UUID key = entry.getKey();
-                int ticketCount = entry.getValue().getFirst().size();
-                int tickingTicketCount = entry.getValue().getSecond().size();
-                LOGGER.info("Allowing {} chunk tickets and {} ticking chunk tickets to be reinstated for entity: {}.", ticketCount, tickingTicketCount, key);
-            }
-        }));
+    private void registerControllers(RegisterTicketControllersEvent event) {
+        event.register(CONTROLLER);
     }
 
     private static class ChunkLoaderBlock extends Block {
@@ -86,7 +87,7 @@ public class ForgeChunkManagerTest {
             super.onPlace(state, worldIn, pos, oldState, isMoving);
             if (worldIn instanceof ServerLevel) {
                 ChunkPos chunkPos = new ChunkPos(pos);
-                ForcedChunkManager.forceChunk((ServerLevel) worldIn, MODID, pos, chunkPos.x, chunkPos.z, true, false);
+                CONTROLLER.forceChunk((ServerLevel) worldIn, pos, chunkPos.x, chunkPos.z, true, true);
             }
         }
 
@@ -95,7 +96,7 @@ public class ForgeChunkManagerTest {
             super.onRemove(state, worldIn, pos, newState, isMoving);
             if (worldIn instanceof ServerLevel && !state.is(newState.getBlock())) {
                 ChunkPos chunkPos = new ChunkPos(pos);
-                ForcedChunkManager.forceChunk((ServerLevel) worldIn, MODID, pos, chunkPos.x, chunkPos.z, false, false);
+                CONTROLLER.forceChunk((ServerLevel) worldIn, pos, chunkPos.x, chunkPos.z, false, true);
             }
         }
     }
